requirements.txt
playwright
openai
python-dotenv
markdown2
pdfkit
playwright-stealth